code/requirements.txt
CairoSVG==2.7.1
fpdf==1.7.2
intervaltree==3.1.0
ipython==8.12.3
matplotlib==3.8.0
numpy
pandas==2.2.3
Pillow==11.0.0
seaborn==0.13.2
